Untested rework of flight.sh

+++ tmutil/flight.sh
@@ -6,10 +6,12 @@ export PATH=:/bin:/usr/bin:/usr/local/bin
 if [ "x$1" != "xteed" ]; then
   touch flight.sh.log
   if [ -w flight.sh.log ]; then
-    $0 teed 2>&1 | tee -a flight.sh.log
+    exec 5>&1
+    $0 teed 4>&5 2>&1 | tee -a flight.sh.log
     exit 0
   fi
   echo "Cannot write to flight.sh.log"
+  exec 4>&1
 fi
 echo "\nRunning flight.sh: \c"; date
 
@@ -35,33 +37,30 @@ export Experiment
 umask g+w
 
 typeset launch_error=""
-typeset Looping=no
 typeset parent_loop=''
-[ -n "$FlightNode" -a -n "$LOOP_ON_FILE" ] && Looping=yes
+typeset script=''
 
-if [ "$Looping" = yes ]; then
-  [ -n "$LOOP_STOP_FILE" ] || LOOP_STOP_FILE=loopstop.txt
+[ -n "$LOOP_STOP_FILE" ] || LOOP_STOP_FILE=loopstop.txt
+[ -n "$LOOP_START_FILE" ] || LOOP_START_FILE=loopstart.txt
+if [ -n "$LOOP_ON_FILE" -a -e "$LOOP_ON_FILE" ]; then
+  echo "`date '+%F %T'`: Invoking reduce for LOOP_ON_FILE"
+  [ -n "$REDUCE_LOG_FILE" ] && REDUCE_LOG_FILE=reducelog.txt
+  exec 1>&4 2>&4 # close pipe to flight.sh.log and reopen stdout/err
+  BEDTIME=yes reduce 2>&1 | tee -a $REDUCE_LOG_FILE
   if [ -e "$LOOP_ON_FILE" ]; then
-    echo "`date '+%F %T'`: Invoking reduce for LOOP_ON_FILE"
-    BEDTIME=yes reduce
-    if [ -e "$LOOP_ON_FILE" ]; then
-      echo "`date '+%F %T'`: reduce did not clear LOOP_ON_FILE '$LOOP_ON_FILE'"
-      rm -f $LOOP_STOP_FILE
-      exec parent
-    fi
-  elif [ -e "$LOOP_STOP_FILE" ]; then
-    echo "`date '+%F %T'`: Looping terminated"
-    rm -f $LOOP_STOP_FILE
+    echo "`date '+%F %T'`: reduce did not clear LOOP_ON_FILE '$LOOP_ON_FILE'"
+    rm -f $LOOP_STOP_FILE $LOOP_START_FILE
     exec parent
-  elif [ -n "$SCRIPT_OVERRIDE" ]; then
-    if [ -e "$SCRIPT_OVERRIDE" ]; then
-      echo "`date '+%F %T'`: Interactive invocation"
-    else
-      echo "`date '+%F %T'`: Bootup invocation"
-      exec parent
-    fi
   fi
-  rm -f $LOOP_STOP_FILE
+  [ -f $LOOP_STOP_FILE ] && mv $LOOP_STOP_FILE $LOOP_START_FILE
+  exit 0 # exit to reestablish tee log
+elif [ -e "$LOOP_STOP_FILE" ]; then
+  echo "`date '+%F %T'`: Looping terminated"
+  rm -f $LOOP_STOP_FILE $LOOP_START_FILE
+  script=/dev/null
+elif [ -e "$LOOP_START_FILE" ]; then
+  script=`cat $LOOP_START_FILE`
+  rm -f $LOOP_START_FILE
 fi
 
 function Launch {
@@ -70,14 +69,15 @@ function Launch {
   [ -n "$launch_error" ] && return 1
   [ -n "$VERBOSE" ] && echo "Launch: $*"
   if { $* & }; then
-    if [ $Looping = yes -a "$name" = "cmd/server" ]; then
-      parent_loop="-q -M $! -t 5"
-      touch $LOOP_STOP_FILE
+    Launch_pid=$!
+    echo "Launch: $Launch_pid $*"
+    if [ "$name" = "DG/cmd" ]; then
+      parent_loop="-q -M $Launch_pid -t 5"
     fi
     if [ "$name" != "-" ]; then
       [ "$name" = "-DC-" ] && name="/var/huarp/run/$Experiment/$!"
       [ "${name#/}" = "$name" ] && name="/dev/huarp/$Experiment/$name"
-      [ -n "$VERBOSE" ] && echo "Launch: Waiting for $!:$name"
+      [ -n "$VERBOSE" ] && echo "Launch: Waiting for $Launch_pid:$name"
       waitfor $name 10 || {
         echo "Launch namewait failure: $*" >&2
         launch_error=yes
@@ -106,14 +106,17 @@ else
 fi
 export TMBINDIR
 
-if [ -n "$SCRIPT_OVERRIDE" -a -r "$SCRIPT_OVERRIDE" ]; then
-  script=`cat $SCRIPT_OVERRIDE`
-  rm $SCRIPT_OVERRIDE
-elif [ -n "$PICKFILE" ]; then
-  script=`cd $TMBINDIR; $PICKFILE`
-else
-  script=${RUNFILE:-runfile.dflt}
+if [ -z "$script" ]; then
+  if [ -n "$SCRIPT_OVERRIDE" -a -r "$SCRIPT_OVERRIDE" ]; then
+    script=`cat $SCRIPT_OVERRIDE`
+    rm $SCRIPT_OVERRIDE
+  elif [ -n "$PICKFILE" ]; then
+    script=`cd $TMBINDIR; $PICKFILE`
+  else
+    script=${RUNFILE:-runfile.dflt}
+  fi
 fi
+  
 [ ${script#/} = "$script" ] && script=$TMBINDIR/$script
 if [ -r $script ]; then
   echo flight.sh: `id`: Experiment=$Experiment script=$script
@@ -121,8 +124,13 @@ if [ -r $script ]; then
 else
   echo flight.sh: Specified script $script not found >&2
 fi
-# [ -z "$launch_error" ] && pick_file -q
 
-typeset qoc
-[ -z "$FlightNode" ] && qoc="-q"
-exec parent $qoc $parent_loop
+if [ -n "$launch_error" ] || jobs -p | grep -q .; then
+  echo $script >$LOOP_STOP_FILE
+else
+  echo "flight.sh: No subprocesses, closing flight.sh.log"
+  exec 1>&4 2>&4
+fi
+
+[ -z "$FlightNode" -a -z "$parent_loop" ] && parent_loop="-q"
+exec parent $parent_loop
